requirements.txt
#Cython
pandas
matplotlib
tqdm
shap
scikit-learn
scipy
#tensorflow
regex
#tensorflow-addons
ipython
numba
opencv-python
#optuna
#pytorch-lightning
seaborn
torch
#tqdm
homcloud 
#alibi_detect[torch]
#torchvision
prompt_toolkit
hydra-core
mlflow
torch-tb-profiler
hydra-optuna-sweeper
#deepchem
#captum
pytest
ngl_view
ase
